requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=3.1.1 # Apache-2.0
sshpubkeys>=2.2.0
pycrypto>=2.6.1
falcon>=1.3.0
sqlalchemy>=1.1.14
castellan>=0.16.0
requests>=2.18.2 # Apache-2.0
python-keystoneclient!=1.8.0,!=2.1.0,>=1.7.0 # Apache-2.0
keystoneauth1>=2.7.0 # Apache-2.0
keystonemiddleware>=4.12.0  # Apache-2.0
oslo.messaging>=5.34.0 # Apache-2.0
oslo.log>=3.36.0 # Apache-2.0
pyramid>=1.9.1 # BSD-derived (http://www.repoze.org/LICENSE.txt)
Paste # MIT
dogpile.cache
python-memcached
python-designateclient
python-neutronclient
python-novaclient
dragonflow
eventlet
vine
python-designateclient
